+++ eras/shelley/impl/testlib/Test/Cardano/Ledger/Shelley/Imp/LedgerSpec.hs
@@ -5,9 +5,9 @@ module Test.Cardano.Ledger.Shelley.Imp.LedgerSpec (
   spec,
 ) where
 
-import Cardano.Ledger.BaseTypes (TxIx (..), inject)
+import Cardano.Ledger.BaseTypes (inject)
 import Cardano.Ledger.Coin (Coin (..))
-import Cardano.Ledger.Core (EraTx (..), EraTxBody (..), EraTxOut (..), coinTxOutL)
+import Cardano.Ledger.Shelley.Core
 import Cardano.Ledger.Shelley.LedgerState (
   esLStateL,
   lsUTxOStateL,
@@ -15,7 +15,7 @@ import Cardano.Ledger.Shelley.LedgerState (
   utxosUtxoL,
  )
 import Cardano.Ledger.Shelley.UTxO (UTxO (..))
-import Cardano.Ledger.TxIn (TxIn (..))
+import Cardano.Ledger.TxIn (mkTxInPartial)
 import qualified Data.Map.Strict as Map
 import qualified Data.Sequence.Strict as SSeq
 import qualified Data.Set as Set
@@ -38,31 +38,30 @@ spec = describe "LEDGER" $ do
     UTxO utxo0 <- getUTxO
     length utxo0 `shouldBe` 1
     let coin1 = Coin 1000
-    txId1 <-
-      submitTx "First transaction" . mkBasicTx $
+    tx1 <-
+      submitTxAnn "First transaction" . mkBasicTx $
         mkBasicTxBody
           & outputsTxBodyL @era
             .~ SSeq.singleton
               (mkBasicTxOut (mkAddr (kpPayment1, kpStaking1)) $ inject coin1)
     UTxO utxo1 <- getUTxO
-    case Map.lookup (TxIn txId1 $ TxIx 1) utxo1 of
-      Just out1 -> do
-        out1 ^. coinTxOutL `shouldBe` coin1
+    case Map.lookup (mkTxInPartial (txIdTx tx1) 1) utxo1 of
+      Just out1 -> out1 ^. coinTxOutL `shouldBe` coin1
       Nothing -> expectationFailure "Could not find the TxOut of the first transaction"
     kpPayment2 <- lookupKeyPair =<< freshKeyHash
     kpStaking2 <- lookupKeyPair =<< freshKeyHash
     let coin2 = Coin 500
-    txId2 <-
-      submitTx "Second transaction" . mkBasicTx $
+    tx2 <-
+      submitTxAnn "Second transaction" . mkBasicTx $
         mkBasicTxBody
           & inputsTxBodyL
             .~ Set.singleton
-              (TxIn txId1 $ TxIx 0)
+              (mkTxInPartial (txIdTx tx1) 0)
           & outputsTxBodyL @era
             .~ SSeq.singleton
               (mkBasicTxOut (mkAddr (kpPayment2, kpStaking2)) $ inject coin2)
     UTxO utxo2 <- getUTxO
-    case Map.lookup (TxIn txId2 $ TxIx 1) utxo2 of
+    case Map.lookup (mkTxInPartial (txIdTx tx2) 1) utxo2 of
       Just out1 -> do
         out1 ^. coinTxOutL `shouldBe` coin2
       Nothing -> expectationFailure "Could not find the TxOut of the second transaction"