+++ libs/cardano-ledger-core/testlib/Test/Cardano/Ledger/Common.hs
@@ -3,6 +3,7 @@ module Test.Cardano.Ledger.Common (
   ledgerTestMain,
   ledgerTestMainWith,
   ledgerHspecConfig,
+  NFData,
 
   -- * Expr
   ToExpr (..),